Boolean working.
