+++ demo/index.html
@@ -52,12 +52,9 @@
           ]
         }]
       },
-      {
-        label: 'Mapping', type: 'object', key: 'mapping'
-      },
-      {
-        label: 'The Blob', type: 'json', key: 'blob'
-      }
+      {label: 'Mapping', type: 'object', key: 'mapping'},
+      {label: 'The Blob', type: 'json', key: 'blob'},
+      {label: 'Do you like cookies?', type: 'boolean', key: 'likesCookies'}
     ];
 
     var formValue = {};
