+++ lib/components/fields/select.js
@@ -9,9 +9,6 @@ boolean values, but it _should_ work for other values.
 'use strict';
 
 var React = require('react/addons');
-var R = React.DOM;
-var _ = require('underscore');
-var cx = React.addons.classSet;
 
 module.exports = React.createClass({
 
@@ -31,16 +28,6 @@ module.exports = React.createClass({
     });
   },
 
-  onChange: function (event) {
-    var choiceValue = event.target.value;
-    var choiceType = choiceValue.substring(0, choiceValue.indexOf(':'));
-    if (choiceType === 'choice') {
-      var choiceIndex = choiceValue.substring(choiceValue.indexOf(':') + 1);
-      choiceIndex = parseInt(choiceIndex);
-      this.onChangeValue(this.state.choices[choiceIndex].value);
-    }
-  },
-
   render: function () {
     return this.renderWithConfig();
   },
@@ -49,62 +36,11 @@ module.exports = React.createClass({
 
     var config = this.props.config;
     var field = this.props.field;
-    var choices = this.state.choices || [];
-
-    var choicesOrLoading;
-
-    if (choices.length === 1 && choices[0].value === '///loading///') {
-      choicesOrLoading = R.div({},
-        'Loading choices...'
-      );
-    } else {
-
-      var value = this.props.value !== undefined ? this.props.value : '';
-
-      choices = choices.map(function (choice, i) {
-        return {
-          choiceValue: 'choice:' + i,
-          value: choice.value,
-          label: choice.label
-        };
-      });
-
-      var valueChoice = _.find(choices, function (choice) {
-        return choice.value === value;
-      });
-
-      if (valueChoice === undefined) {
-
-        var label = value;
-        if (!_.isString(value)) {
-          label = JSON.stringify(value);
-        }
-        valueChoice = {
-          choiceValue: 'value:',
-          value: value,
-          label: label
-        };
-        choices = [valueChoice].concat(choices);
-      }
-
-      choicesOrLoading = R.select({
-        className: cx(this.props.classes),
-        onChange: this.onChange,
-        value: valueChoice.choiceValue,
-        onFocus: this.onFocus,
-        onBlur: this.onBlur
-      },
-        choices.map(function (choice, i) {
-          return R.option({
-            key: i,
-            value: choice.choiceValue
-          }, choice.label);
-        }.bind(this))
-      );
-    }
 
     return config.createElement('field', {
       config: config, field: field, plain: this.props.plain
-    }, choicesOrLoading);
+    }, config.createElement('select-value', {
+      choices: this.state.choices, value: this.props.value, onChange: this.onChangeValue
+    }));
   }
 });
