+++ lib/components/helpers/select-value.js
@@ -1,0 +1,86 @@
+'use strict';
+
+var React = require('react/addons');
+var R = React.DOM;
+var _ = require('underscore');
+var cx = React.addons.classSet;
+
+module.exports = React.createClass({
+
+  displayName: 'SelectValue',
+
+  mixins: [require('../../mixins/config')],
+
+  onChange: function (event) {
+    var choiceValue = event.target.value;
+    var choiceType = choiceValue.substring(0, choiceValue.indexOf(':'));
+    if (choiceType === 'choice') {
+      var choiceIndex = choiceValue.substring(choiceValue.indexOf(':') + 1);
+      choiceIndex = parseInt(choiceIndex);
+      this.props.onChange(this.props.choices[choiceIndex].value);
+    }
+  },
+
+  render: function () {
+    return this.renderWithConfig();
+  },
+
+  renderDefault: function () {
+
+    var choices = this.props.choices || [];
+
+    var choicesOrLoading;
+
+    if (choices.length === 1 && choices[0].value === '///loading///') {
+      choicesOrLoading = R.div({},
+        'Loading choices...'
+      );
+    } else {
+
+      var value = this.props.value !== undefined ? this.props.value : '';
+
+      choices = choices.map(function (choice, i) {
+        return {
+          choiceValue: 'choice:' + i,
+          value: choice.value,
+          label: choice.label
+        };
+      });
+
+      var valueChoice = _.find(choices, function (choice) {
+        return choice.value === value;
+      });
+
+      if (valueChoice === undefined) {
+
+        var label = value;
+        if (!_.isString(value)) {
+          label = JSON.stringify(value);
+        }
+        valueChoice = {
+          choiceValue: 'value:',
+          value: value,
+          label: label
+        };
+        choices = [valueChoice].concat(choices);
+      }
+
+      choicesOrLoading = R.select({
+        className: cx(this.props.classes),
+        onChange: this.onChange,
+        value: valueChoice.choiceValue,
+        onFocus: this.onFocus,
+        onBlur: this.onBlur
+      },
+      choices.map(function (choice, i) {
+        return R.option({
+          key: i,
+          value: choice.choiceValue
+        }, choice.label);
+      }.bind(this))
+    );
+  }
+
+  return choicesOrLoading;
+}
+});
