+++ lib/default-config.js
@@ -17,6 +17,10 @@ var createArray = function () {
   return [];
 };
 
+var createBoolean = function () {
+  return false;
+};
+
 module.exports = {
 
   // Field element factories
@@ -24,6 +28,7 @@ module.exports = {
   createElement_Text: React.createFactory(require('./components/fields/text')),
   createElement_Unicode: React.createFactory(require('./components/fields/unicode')),
   createElement_Select: React.createFactory(require('./components/fields/select')),
+  createElement_Boolean: React.createFactory(require('./components/fields/boolean')),
   createElement_PrettyTextarea: React.createFactory(require('./components/fields/pretty-textarea')),
   createElement_List: React.createFactory(require('./components/fields/list')),
   createElement_Object: React.createFactory(require('./components/fields/object')),
@@ -50,6 +55,7 @@ module.exports = {
   createElement_ObjectItemValue: React.createFactory(require('./components/helpers/object-item-value')),
   createElement_ObjectItemKey: React.createFactory(require('./components/helpers/object-item-key')),
   createElement_ObjectItem: React.createFactory(require('./components/helpers/object-item')),
+  createElement_SelectValue: React.createFactory(require('./components/helpers/select-value')),
 
   // Field default values
   defaultValue_Fields: createObject,
@@ -59,6 +65,7 @@ module.exports = {
   defaultValue_List: createArray,
   defaultValue_Object: createObject,
   defaultValue_Json: createObject,
+  defaultValue_Boolean: createBoolean,
 
   hasElementFactory: function (name) {
     var config = this;
@@ -123,6 +130,7 @@ module.exports = {
   },
 
   alias_Dict: 'Object',
+  alias_Bool: 'Boolean',
 
   // Field helpers
   fieldTypeName: function (field) {
